fix: 6개 버그 수정 - 컴파일 에러·버퍼 오버플로·SEQ 혼용 등
[CRITICAL] s_main.c: cc1101_woke_flag 미정의 변수 제거
  - wake_flags의 Slave_IsWakeFlagSet() 호출로 교체
  - 두 번째 조건의 반전 로직 수정 (스트림 명령 수신 경로)

[CRITICAL] s_main.c & m_main.c: STREAM_LOGICAL_DATA_MAX 17→20
  - logical[] 배열에 4B 헤더 + 16B sensor_data = 20B 필요
  - 17 유지 시 스택 버퍼 오버플로 후 항상 false 반환

[HIGH] s_main.c: resp->seq = req->seq → g_seq++
  - 커맨드 응답이 Master의 seq를 에코하면 스트림 DATA의
    slave g_seq와 공간이 혼용되어 anti-replay 검사 실패

[MEDIUM] m_main.c: Master_WriteToImplant 리턴 코드 -8 중복
  - FCF 타입 검사(-8)와 DeviceID 검사(-8) 구분 → -9로 수정

[LOW] m_main.c: 주석 끝에 달라붙은 중복 if 체크 제거

[DOC] protocol.h: "Maximum LEN field = 63" → 61 (= PKTLEN)

[WARN] stm32u575zi_UART.c: const* → non-const HAL 캐스트 명시

+++ app/stm32u575zi_UART.c
@@ -25,7 +25,7 @@ HAL_StatusTypeDef STM32U575ZI_UART_SendBytes(UART_HandleTypeDef *huart,
         return HAL_ERROR;
     }
 
-    return HAL_UART_Transmit(huart, (uint8_t *)data, len, STM32U575ZI_UART_TX_TIMEOUT_MS);
+    return HAL_UART_Transmit(huart, (uint8_t *)(uintptr_t)data, len, STM32U575ZI_UART_TX_TIMEOUT_MS);
 }
 
 HAL_StatusTypeDef STM32U575ZI_UART_SendString(UART_HandleTypeDef *huart,
@@ -47,7 +47,7 @@ HAL_StatusTypeDef STM32U575ZI_UART_SendString(UART_HandleTypeDef *huart,
     }
 
     return HAL_UART_Transmit(huart,
-                             (uint8_t *)str,
+                             (uint8_t *)(uintptr_t)str,
                              (uint16_t)len,
                              STM32U575ZI_UART_TX_TIMEOUT_MS);
 }
